+++ pyproject.toml
@@ -42,3 +42,7 @@ ignore_missing_imports = true
 [[tool.mypy.overrides]]
 module = ["ujson.*"]
 ignore_missing_imports = true
+
+[tool.pyright]
+include = ["src"]
+pythonVersion = "3.8"
