+++ src/dask_awkward/tests/test_core.py
@@ -4,11 +4,10 @@ import pytest
 
 import dask_awkward as dak
 import dask_awkward.core as dakc
-from dask_awkward.testutils import assert_eq
-
-from .helpers import (  # noqa: F401
+from dask_awkward.testutils import (  # noqa: F401
     _lazyrecord,
     _lazyrecords,
+    assert_eq,
     line_delim_records_file,
     load_records_eager,
     load_records_lazy,
