+++ src/dask_awkward/tests/test_reducers.py
@@ -4,9 +4,8 @@ import awkward._v2 as ak
 import pytest
 
 import dask_awkward as dak
-from dask_awkward.testutils import assert_eq
-
-from .helpers import (  # noqa: F401
+from dask_awkward.testutils import (  # noqa: F401
+    assert_eq,
     line_delim_records_file,
     load_records_eager,
     load_records_lazy,
