+++ src/dask_awkward/tests/test_ufunc.py
@@ -3,9 +3,8 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-from dask_awkward.testutils import assert_eq
-
-from .helpers import (  # noqa: F401
+from dask_awkward.testutils import (  # noqa: F401
+    assert_eq,
     line_delim_records_file,
     load_records_eager,
     load_records_lazy,
